requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.28
pandas-ta>=0.3.14
scikit-learn>=1.3.0
xgboost>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
lightgbm